+++ src/main/java/com/sandbox/domain/todos/repository/TodoDao.java
@@ -1,4 +1,4 @@
-package com.sandbox.domain.todos.dao;
+package com.sandbox.domain.todos.repository;
 
 import com.sandbox.domain.todos.entity.Todo;
 import org.springframework.data.jpa.repository.JpaRepository;
